Drop redundant uint8(i) casts when assigning to RaidTargetUpdateAll::IconEntry::iconId
Agent-Logs-Url: https://github.com/0blu/vmangos-core/sessions/6d266702-c725-4f9c-b629-8294c36f4237

+++ src/game/Group/Group.cpp
@@ -1361,7 +1361,7 @@ void Group::SendTargetIconList(WorldSession* session)
             continue;
 
         WorldPackets::Group::RaidTargetUpdateAll::IconEntry entry;
-        entry.iconId = uint8(i);
+        entry.iconId = i;
         entry.targetGuid = m_targetIcons[i];
         packet->icons.push_back(entry);
     }
@@ -1386,7 +1386,7 @@ void Group::SendUpdate()
                 markedTargets = std::make_shared<WorldPackets::Group::RaidTargetUpdateAll>();
 
             WorldPackets::Group::RaidTargetUpdateAll::IconEntry entry;
-            entry.iconId = uint8(i);
+            entry.iconId = i;
             entry.targetGuid = m_targetIcons[i];
             markedTargets->icons.push_back(entry);
         }
